Refactor image tab
