+++ apps/studio/src/routes/editor/LayersPanel/ImageTab.tsx/DeleteModal.tsx
@@ -1,0 +1,49 @@
+import {
+    AlertDialog,
+    AlertDialogContent,
+    AlertDialogDescription,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogTitle,
+} from '@onlook/ui/alert-dialog';
+import { Button } from '@onlook/ui/button';
+
+export default function DeleteImageModal({
+    onDelete,
+    isOpen,
+    toggleOpen,
+}: {
+    onDelete: () => void;
+    isOpen: boolean;
+    toggleOpen: () => void;
+}) {
+    const handleDelete = () => {
+        onDelete();
+        toggleOpen();
+    };
+
+    return (
+        <AlertDialog open={isOpen} onOpenChange={toggleOpen}>
+            <AlertDialogContent>
+                <AlertDialogHeader>
+                    <AlertDialogTitle>{'Delete this image?'}</AlertDialogTitle>
+                    <AlertDialogDescription>
+                        {"This will delete the image from the project. You can't undo this action."}
+                    </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                    <Button variant={'ghost'} onClick={toggleOpen}>
+                        Cancel
+                    </Button>
+                    <Button
+                        variant={'destructive'}
+                        className="rounded-md text-sm"
+                        onClick={handleDelete}
+                    >
+                        Delete
+                    </Button>
+                </AlertDialogFooter>
+            </AlertDialogContent>
+        </AlertDialog>
+    );
+}
